+++ CHANGELOG.md
@@ -18,7 +18,6 @@ Note: a potentially breaking change for docker users: \the default user is now "
  - Authentication enhancements
    - LDAP login support with OIDC feature parity. (#591)
    - userGroup for OIDC and LDAP, only users in a group will get access. #1964
-   - 2 minute upstream login check for LDAP and OIDC. So if a user logs out of the OIDC/LDAP provider, their filebrowser session will expire within 2 minutes
  - Enhanced thumbnail and item previews
    - Added ability to show "motion preview" for folders with multiple child items that have previews. cycles through the first 4 images.
    - support for reading embedded images from raw image or heic/heif files (#215)
@@ -51,6 +50,8 @@ Note: a potentially breaking change for docker users: \the default user is now "
    - `/api/preview` has been removed and replaced with `/api/resources/preview`
    - `/api/onlyoffice` have been replaced with `/api/office`
    - `/api/shares` has been moved to `/api/share/list`
+   - `/api/auth/tokens` has been moved to `/api/auth/token/list` and `/api/auth/token` added to get specific token info
+   - `PUT /api/token` has been moved to `POST /api/token`
    - `/public/api/shareinfo` has been moved to `/public/api/share/info`
    - `POST /resources/bulk/delete` api has been moved to `DELETE /resources/bulk` (#1984)